Extract separate observer package
